unify launch geometry, split off cooperative-reduce body rewrite
Swap pass order: launch geometry decides Tile.axes first for all body
shapes (matmul, cooperative reduce, pointwise); cooperative reduce now
runs second as pure body rewrite (Loop → StridedLoop, Combine insertion,
naked-axis epilogue wrap).

- 004_launch_geometry (renamed from 005_blockify_launch): one pass
  owning launch geometry. Matmul fork over (BN, BM); cooperative-reduce
  election adds synthetic t=THREAD + BLOCK-binds outputs; pointwise
  deterministic partition. Reduce-viability gate (extent ≥ WARP_SIZE,
  valid independent Accums) lives here.
- 005_cooperative_reduce (renamed, slimmed from 004_cooperative_reduce):
  pure body rewrite triggered by `len(thread_axes) == 1` + reduce Loop.

Unblocks hybrid shapes (per-CTA output tile + cooperative thread axis,
needed for flash-attention-style kernels) without touching 006-016
downstream. All 609 compiler tests pass; perf unchanged on softmax /
matmul / rmsnorm smoke kernels.
